refactor to try and simplfy a bit
