+++ packages/scenes/src/querying/layers/annotations/filterAnnotations.test.ts
@@ -1,9 +1,8 @@
 import { arrayToDataFrame } from '@grafana/data';
-import { of } from 'rxjs';
-import { filterAnnotationsOperator } from './filterAnnotationsOperator';
+import { filterAnnotations } from './filterAnnotations';
 
-describe('filterAnnotationsOperator', () => {
-  test('multiple frames', (done) => {
+describe('filterAnnotations', () => {
+  test('multiple frames', () => {
     const events1 = [
       {
         source: {
@@ -45,15 +44,10 @@ describe('filterAnnotationsOperator', () => {
     const df1 = arrayToDataFrame(events1);
     const df2 = arrayToDataFrame(events2);
 
-    const filter = filterAnnotationsOperator({ panelId: 2 })({
-      interpolate: () => '',
-    });
+    const result = filterAnnotations([df1, df2], { panelId: 2 });
 
-    filter(of([df1, df2])).subscribe((frames) => {
-      expect(frames.length).toBe(2);
-      expect(frames[0].length).toBe(0);
-      expect(frames[1].length).toBe(2);
-      done();
-    });
+    expect(result.length).toBe(2);
+    expect(result[0].length).toBe(0);
+    expect(result[1].length).toBe(2);
   });
 });
